+++ airflow/contrib/hooks/aws_firehose_hook.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 #
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
@@ -16,46 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""
-This module contains AWS Firehose hook
-"""
-from airflow.contrib.hooks.aws_hook import AwsHook
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.kinesis`."""
 
+import warnings
 
-class AwsFirehoseHook(AwsHook):
-    """
-    Interact with AWS Kinesis Firehose.
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.hooks.kinesis import AwsFirehoseHook  # noqa
 
-    :param delivery_stream: Name of the delivery stream
-    :type delivery_stream: str
-    :param region_name: AWS region name (example: us-east-1)
-    :type region_name: str
-    """
-
-    def __init__(self, delivery_stream, region_name=None, *args, **kwargs):
-        self.delivery_stream = delivery_stream
-        self.region_name = region_name
-        self.conn = None
-        super().__init__(*args, **kwargs)
-
-    def get_conn(self):
-        """
-        Returns AwsHook connection object.
-        """
-
-        self.conn = self.get_client_type('firehose', self.region_name)
-        return self.conn
-
-    def put_records(self, records):
-        """
-        Write batch records to Kinesis Firehose
-        """
-
-        firehose_conn = self.get_conn()
-
-        response = firehose_conn.put_record_batch(
-            DeliveryStreamName=self.delivery_stream,
-            Records=records
-        )
-
-        return response
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.kinesis`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
